Change AmpInfoRecord saturation level to double
It turns out that IsrTask masks saturated pixels after the image
is converted to floating point. Furthermore, the task does no masking
if the saturation level is `nan`. However, the saturation level
was an integer, so I changed the saturation level to a float
to match the task code.

Also document `nan` to mean no level in the help for the field.

+++ include/lsst/afw/table/AmpInfo.h
@@ -100,8 +100,8 @@ public:
     double getReadNoise() const;
     void setReadNoise(double readNoise); ///< amplifier read noise, in e-
 
-    int getSaturation() const;
-    void setSaturation(int saturation); ///< saturation value, in ADU
+    double getSaturation() const;
+    void setSaturation(double saturation); ///< saturation value, in ADU
 
     ReadoutCorner getReadoutCorner() const;
     void setReadoutCorner(ReadoutCorner val); ///< set readout corner
@@ -204,7 +204,7 @@ public:
     static PointKey<int> getBBoxExtentKey() { return getMinimalSchema().bboxExtent; }
     static Key<double> getGainKey() { return getMinimalSchema().gain; }
     static Key<double> getReadNoiseKey() { return getMinimalSchema().readNoise; }
-    static Key<int> getSaturationKey() { return getMinimalSchema().saturation; }
+    static Key<double> getSaturationKey() { return getMinimalSchema().saturation; }
     static Key<int> getReadoutCornerKey() { return getMinimalSchema().readoutCorner; }
     static Key< Array<double> > getLinearityCoeffsKey() { return getMinimalSchema().linearityCoeffs; }
     static Key<std::string> getLinearityTypeKey() { return getMinimalSchema().linearityType; }
@@ -256,7 +256,7 @@ private:
         PointKey<int> bboxExtent;
         Key<double> gain;
         Key<double> readNoise;
-        Key<int> saturation;
+        Key<double> saturation;
         Key<int> readoutCorner;
         Key< Array<double> > linearityCoeffs;
         Key<std::string> linearityType;
